reorganise server.go
