+++ kv/tikv/scheduler.go
@@ -1,0 +1,32 @@
+package tikv
+
+import (
+	"github.com/pingcap-incubator/tinykv/kv/tikv/storage/kvstore"
+	"github.com/pingcap-incubator/tinykv/proto/pkg/kvrpcpb"
+)
+
+// Interfaces used with the scheduler and related code in storage.
+
+// Scheduler takes Commands and runs them asynchronously. It is up to implementations to decide the scheduling policy.
+type Scheduler interface {
+	// Run executes a command asynchronously and returns a channel over which the result of the command execution is
+	// sent.
+	Run(Command) <-chan RespResult
+	Stop()
+}
+
+// Command is an abstraction which covers the process from receiving a request from gRPC to returning a response.
+// That process is driven by a Scheduler.
+type Command interface {
+	// BuildTxn is for building writes into an write using an mvcc transaction. Commands can also make non-transactional
+	// reads and writes using txn. Returning without modifying txn means that no transaction will be executed.
+	BuildTxn(txn *kvstore.Txn) error
+	Context() *kvrpcpb.Context
+	// Response builds a success response to return to the client.
+	Response() interface{}
+	// HandleError gives the command an opportunity to handle errors generated at any stage in processing. If the command
+	// handles the error, it should return a success response to return to the user. If nil is returned then an error
+	// will be generated by the scheduler. Note that there is no way to continue processing after an error, even if the
+	// command handles the error.
+	HandleError(err error) interface{}
+}
