+++ kv/tikv/server.go
@@ -9,7 +9,6 @@ import (
 	"github.com/pingcap-incubator/tinykv/kv/tikv/dbreader"
 	"github.com/pingcap-incubator/tinykv/kv/tikv/inner_server"
 	"github.com/pingcap-incubator/tinykv/kv/tikv/storage/commands"
-	"github.com/pingcap-incubator/tinykv/kv/tikv/storage/kvstore"
 	"github.com/pingcap-incubator/tinykv/proto/pkg/coprocessor"
 	"github.com/pingcap-incubator/tinykv/proto/pkg/errorpb"
 	"github.com/pingcap-incubator/tinykv/proto/pkg/kvrpcpb"
@@ -35,130 +34,6 @@ type InnerServer interface {
 	Reader(ctx *kvrpcpb.Context) (dbreader.DBReader, error)
 	Raft(stream tikvpb.Tikv_RaftServer) error
 	Snapshot(stream tikvpb.Tikv_SnapshotServer) error
-}
-
-// Scheduler takes Commands and runs them asynchronously. It is up to implementations to decide the scheduling policy.
-type Scheduler interface {
-	// Run executes a command asynchronously and returns a channel over which the result of the command execution is
-	// sent.
-	Run(Command) <-chan RespResult
-	Stop()
-}
-
-// RespResult is a 'generic' result type for responses. It is used to return a Response/error pair over channels where
-// we can't use Go's multiple return values.
-type RespResult struct {
-	Response interface{}
-	Err      error
-}
-
-func RespOk(resp interface{}) RespResult {
-	return RespResult{
-		Response: resp,
-		Err:      nil,
-	}
-}
-
-func RespErr(err error) RespResult {
-	return RespResult{
-		Response: nil,
-		Err:      err,
-	}
-}
-
-func (rr *RespResult) getResponse() *kvrpcpb.GetResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.GetResponse)
-}
-func (rr *RespResult) prewriteResponse() *kvrpcpb.PrewriteResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.PrewriteResponse)
-}
-func (rr *RespResult) scan() *kvrpcpb.ScanResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.ScanResponse)
-}
-func (rr *RespResult) checkTxnStatus() *kvrpcpb.CheckTxnStatusResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.CheckTxnStatusResponse)
-}
-func (rr *RespResult) commit() *kvrpcpb.CommitResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.CommitResponse)
-}
-func (rr *RespResult) cleanup() *kvrpcpb.CleanupResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.CleanupResponse)
-}
-func (rr *RespResult) batchRollback() *kvrpcpb.BatchRollbackResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.BatchRollbackResponse)
-}
-func (rr *RespResult) scanLock() *kvrpcpb.ScanLockResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.ScanLockResponse)
-}
-func (rr *RespResult) resolveLock() *kvrpcpb.ResolveLockResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.ResolveLockResponse)
-}
-func (rr *RespResult) rawGetResponse() *kvrpcpb.RawGetResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.RawGetResponse)
-}
-func (rr *RespResult) rawPutResponse() *kvrpcpb.RawPutResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.RawPutResponse)
-}
-func (rr *RespResult) rawDeleteResponse() *kvrpcpb.RawDeleteResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.RawDeleteResponse)
-}
-func (rr *RespResult) rawScanResponse() *kvrpcpb.RawScanResponse {
-	if rr.Response == nil {
-		return nil
-	}
-	return rr.Response.(*kvrpcpb.RawScanResponse)
-}
-
-// Command is an abstraction which covers the process from receiving a request from gRPC to returning a response.
-// That process is driven by a Scheduler.
-type Command interface {
-	// BuildTxn is for building writes into an write using an mvcc transaction. Commands can also make non-transactional
-	// reads and writes using txn. Returning without modifying txn means that no transaction will be executed.
-	BuildTxn(txn *kvstore.Txn) error
-	Context() *kvrpcpb.Context
-	// Response builds a success response to return to the client.
-	Response() interface{}
-	// HandleError gives the command an opportunity to handle errors generated at any stage in processing. If the command
-	// handles the error, it should return a success response to return to the user. If nil is returned then an error
-	// will be generated by the scheduler. Note that there is no way to continue processing after an error, even if the
-	// command handles the error.
-	HandleError(err error) interface{}
 }
 
 func NewServer(innerServer InnerServer, scheduler Scheduler) *Server {
